Address review feedback from #4328
- Write parquet instead of JSONL in all rollout transforms
- Add zephyr counters for kept/dropped rows
- Replace map+filter with flat_map returning [] or [doc]
- Extract shared load_parquet_batched and strip_think_tags
