+++ lib/marin/src/marin/datakit/download/rollout_transforms.py
@@ -1,0 +1,24 @@
+# Copyright The Marin Authors
+# SPDX-License-Identifier: Apache-2.0
+
+"""Shared transform helpers for rollout dataset pipelines."""
+
+from collections.abc import Iterator
+
+import pyarrow.parquet as pq
+from rigging.filesystem import open_url
+
+
+def load_parquet_batched(path: str) -> Iterator[dict]:
+    """Read parquet via iter_batches to avoid OOM on large nested-struct columns."""
+    with open_url(path, "rb") as f:
+        pf = pq.ParquetFile(f)
+        for batch in pf.iter_batches(batch_size=16):
+            rows = batch.to_pydict()
+            n = len(next(iter(rows.values())))
+            for i in range(n):
+                yield {k: rows[k][i] for k in rows}
+
+
+def strip_think_tags(text: str) -> str:
+    return text.replace("<think>", "").replace("</think>", "").strip()
