+++ lib/marin/src/marin/datakit/download/swe_rebench_openhands.py
@@ -9,14 +9,12 @@ a resolved flag indicating whether the trajectory solved the issue.
 """
 
 import hashlib
-from collections.abc import Iterator
 
-import pyarrow.parquet as pq
 from fray.v2 import ResourceConfig
-from rigging.filesystem import open_url
-from zephyr import Dataset, ZephyrContext
+from zephyr import Dataset, ZephyrContext, counters
 
 from marin.datakit.download.huggingface import download_hf_step
+from marin.datakit.download.rollout_transforms import load_parquet_batched
 from marin.execution.step_spec import StepSpec
 
 HF_DATASET_ID = "nebius/SWE-rebench-openhands-trajectories"
@@ -37,40 +35,32 @@ def render_message(msg: dict) -> str:
     return f"<{role}>\n{content}\n</{role}>"
 
 
-def row_to_doc(row: dict) -> dict | None:
+def row_to_doc(row: dict) -> list[dict]:
     trajectory = row.get("trajectory")
     if not trajectory:
-        return None
+        counters.increment("swe_rebench_openhands/dropped")
+        return []
 
     tag = resolved_to_tag(row.get("resolved"))
     rendered = "\n\n".join(render_message(m) for m in trajectory)
     text = f"{tag}\n\n{rendered}" if tag else rendered
 
-    return {
-        "id": hashlib.sha256(text.encode("utf-8")).hexdigest(),
-        "text": text,
-        "source": "nebius/SWE-rebench-openhands-trajectories",
-    }
-
-
-def load_parquet_batched(path: str) -> Iterator[dict]:
-    """Read parquet via iter_batches to avoid OOM on large nested-struct columns."""
-    with open_url(path, "rb") as f:
-        pf = pq.ParquetFile(f)
-        for batch in pf.iter_batches(batch_size=16):
-            rows = batch.to_pydict()
-            n = len(next(iter(rows.values())))
-            for i in range(n):
-                yield {k: rows[k][i] for k in rows}
+    counters.increment("swe_rebench_openhands/kept")
+    return [
+        {
+            "id": hashlib.sha256(text.encode("utf-8")).hexdigest(),
+            "text": text,
+            "source": "nebius/SWE-rebench-openhands-trajectories",
+        }
+    ]
 
 
 def transform(input_path: str, output_path: str) -> None:
     pipeline = (
         Dataset.from_files(f"{input_path}/**/*.parquet")
         .flat_map(load_parquet_batched)
-        .map(row_to_doc)
-        .filter(lambda r: r is not None)
-        .write_jsonl(f"{output_path}/data-{{shard:05d}}-of-{{total:05d}}.jsonl.gz", skip_existing=True)
+        .flat_map(row_to_doc)
+        .write_parquet(f"{output_path}/data-{{shard:05d}}-of-{{total:05d}}.parquet", skip_existing=True)
     )
     ctx = ZephyrContext(name="swe-rebench-openhands-transform", resources=ResourceConfig(cpu=1, ram="32g"))
     list(ctx.execute(pipeline))
